feat: Add media volume to docker-compose.yml and update product and category interfaces
The `docker-compose.yml` file was modified to include a new volume mapping for the `ecommerce_backend/media` directory. This will allow media files to be persisted between container restarts.

Additionally, the `icategory.interface.ts` and `iproduct.interface.ts` files were updated to include a new `id` field in the `ICategory` and `IProduct` interfaces respectively. This will provide a unique identifier for each category and product.

These changes will improve the handling of media files and provide better data structure for categories and products.