feat: Update media handling and product interfaces
The `urls.py` file was modified to include a new regular expression pattern for media files in the `media` URL path. This change allows for the inclusion of the `.avif` file extension in addition to `.jpg`, `.jpeg`, `.png`, `.gif`, and `.webp` extensions.

In the `login.component.ts` file, a console log statement was commented out to improve code readability.

The `list-products.component.ts` file was updated to remove the unnecessary `response` parameter in the `next` callback function.

In the `views.py` file, the `product` field was renamed to `name` in the `list_carts` view function.

The `list-products.component.scss` file was modified to include styles for an image container and a responsive image.

The `list-products.component.html` file was updated to wrap the product image in a container with a fixed height and width, and to apply a responsive image class to the image.

In the `cart.service.ts` file, the `manageCart` function was updated to add the new cart payload to the existing list of carts.

These changes improve media handling and update the product interfaces for better data structure and code readability.